Some english language corrections.
I've corrected a spelling error, but I don't know what "IP-Address to check the Attacks" means. Does it mean "check the IP address of the source of the attack"?

Also, I'm not sure of the meaning of "issue attacks", perhaps it means "report attacks"? Can we use the API to report attacks? Perhaps that line should read "The API can currently report attacks and issue reports per user, server ID or ip-address."

and "The *appeal* is therefore…". Perhaps "The application" or "The API"?